# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sun Proactive AI — Autonomous Social Impact Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all tasks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, active, completed, cancelled]
        - name: category
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"

    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/publish:
    post:
      operationId: publishTask
      tags: [tasks]
      summary: Publish a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/apply:
    post:
      operationId: applyToTask
      tags: [tasks]
      summary: Volunteer applies to a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /tasks/{id}/applications:
    get:
      operationId: getTaskApplications
      tags: [tasks]
      summary: Get applications for a task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /tasks/{id}/feedback:
    post:
      operationId: submitTaskFeedback
      tags: [tasks]
      summary: Admin submits feedback for a completed task
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "200":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
  /tasks/dashboard/summary:
    get:
      operationId: getTaskDashboardSummary
      tags: [tasks]
      summary: Get dashboard summary stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /tasks/dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [tasks]
      summary: Get recent platform activity feed
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # Volunteers
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        location:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        date:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in hours
        requiredPeople:
          type: integer
        skillsRequired:
          type: array
          items:
            type: string
        ageRequirements:
          type: string
          nullable: true
        category:
          type: string
        status:
          type: string
          enum: [draft, published, active, completed, cancelled]
        qualityScore:
          type: integer
          nullable: true
        successPrediction:
          type: integer
          nullable: true
        createdById:
          type: integer
        applicantsCount:
          type: integer
        approvedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, location, date, duration, requiredPeople, skillsRequired, category, status, createdById, applicantsCount, approvedCount, createdAt]
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/Task"
        - type: object
          properties:
            applications:
              type: array
              items:
                $ref: "#/components/schemas/Application"
            teamRoles:
              $ref: "#/components/schemas/TeamRoles"
              nullable: true
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        date:
          type: string
          format: date-time
        duration:
          type: integer
        requiredPeople:
          type: integer
        skillsRequired:
          type: array
          items:
            type: string
        ageRequirements:
          type: string
          nullable: true
        category:
          type: string
      required: [title, description, location, date, duration, requiredPeople, skillsRequired, category]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        date:
          type: string
          format: date-time
        duration:
          type: integer
        requiredPeople:
          type: integer
        skillsRequired:
          type: array
          items:
            type: string
        category:
          type: string
        status:
          type: string
          enum: [draft, published, active, completed, cancelled]
    Application:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        volunteerId:
          type: integer
        volunteerName:
          type: string
        volunteerEmail:
          type: string
        volunteerSkills:
          type: array
          items:
            type: string
        matchScore:
          type: integer
          nullable: true
        role:
          type: string
          nullable: true
          enum: [leader, executor, communicator, null]
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        completionPhotoUrl:
          type: string
          nullable: true
        verificationStatus:
          type: string
          nullable: true
          enum: [pending, approved, rejected, null]
        verificationReason:
          type: string
          nullable: true
        hoursAwarded:
          type: number
          nullable: true
        adminRating:
          type: number
          nullable: true
        adminComment:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
      required: [id, taskId, volunteerId, volunteerName, volunteerEmail, volunteerSkills, status, appliedAt]
    TeamRoles:
      type: object
      properties:
        leader:
          $ref: "#/components/schemas/TeamMember"
          nullable: true
        executors:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        communicators:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
      required: [executors, communicators]
    TeamMember:
      type: object
      properties:
        volunteerId:
          type: integer
        name:
          type: string
        role:
          type: string
        matchScore:
          type: integer
        reason:
          type: string
      required: [volunteerId, name, role, matchScore, reason]
    FeedbackRequest:
      type: object
      properties:
        applicationId:
          type: integer
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
        hoursAwarded:
          type: number
      required: [applicationId, rating, hoursAwarded]
    FeedbackResponse:
      type: object
      properties:
        success:
          type: boolean
        hoursAwarded:
          type: number
        newRating:
          type: number
        xpGained:
          type: integer
        newBadges:
          type: array
          items:
            type: string
        newLevel:
          type: integer
          nullable: true
      required: [success, hoursAwarded, newRating, xpGained, newBadges]
    DashboardSummary:
      type: object
      properties:
        totalTasks:
          type: integer
        activeTasks:
          type: integer
        completedTasks:
          type: integer
        totalVolunteers:
          type: integer
        totalHoursContributed:
          type: number
        avgSuccessRate:
          type: number
        tasksByCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required: [category, count]
        recentApplicationsCount:
          type: integer
        pendingVerifications:
          type: integer
      required: [totalTasks, activeTasks, completedTasks, totalVolunteers, totalHoursContributed, avgSuccessRate, tasksByCategory, recentApplicationsCount, pendingVerifications]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [task_created, task_completed, volunteer_joined, badge_earned, task_published]
        title:
          type: string
        description:
          type: string
        actorName:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [id, type, title, description, actorName, timestamp]
...