# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sun Proactive AI — Autonomous Social Impact Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryTasksByCategoryItem } from "./dashboardSummaryTasksByCategoryItem";

export interface DashboardSummary {
  totalTasks: number;
  activeTasks: number;
  completedTasks: number;
  totalVolunteers: number;
  totalHoursContributed: number;
  avgSuccessRate: number;
  tasksByCategory: DashboardSummaryTasksByCategoryItem[];
  recentApplicationsCount: number;
  pendingVerifications: number;
}