# lib/api-zod/src/generated/types/teamRoles.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sun Proactive AI — Autonomous Social Impact Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMember } from "./teamMember";

export interface TeamRoles {
  leader?: TeamMember | null;
  executors: TeamMember[];
  communicators: TeamMember[];
}